Move: image 폴더 변경 #81
