+++ .gitignore
@@ -20,6 +20,7 @@ src/main/resources/static/docs/
 AuthKey_D5ZQTHUQ4K.p8
 AuthKey_ZJ4VQZNW7D.p8
 firebase-adminsdk.json
+gcp-img-role.json
 
 # ignore properties file
 application-COMMON.properties
