+++ src/main/java/com/apps/pochak/post/service/PostService.java
@@ -9,7 +9,7 @@ import com.apps.pochak.comment.domain.Comment;
 import com.apps.pochak.comment.domain.repository.CommentRepository;
 import com.apps.pochak.follow.domain.repository.FollowRepository;
 import com.apps.pochak.global.api_payload.exception.GeneralException;
-import com.apps.pochak.global.s3.S3Service;
+import com.apps.pochak.global.image.S3Service;
 import com.apps.pochak.like.domain.repository.LikeRepository;
 import com.apps.pochak.member.domain.Member;
 import com.apps.pochak.member.domain.repository.MemberRepository;
@@ -32,7 +32,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static com.apps.pochak.global.api_payload.code.status.ErrorStatus.*;
-import static com.apps.pochak.global.s3.DirName.POST;
+import static com.apps.pochak.global.image.DirName.POST;
 
 @Service
 @Transactional
